requirements.txt
dj-database-url==0.4.1
Django==1.9.8
django-allauth==0.26.1
django-widget-tweaks==1.4.1
gunicorn==19.6.0
oauthlib==1.1.2
Pillow==3.3.0
psycopg2==2.6.2
python-http-client==1.2.3
python-openid==2.2.5
requests==2.11.0
requests-oauthlib==0.6.2
sendgrid==2.2.1
sendgrid-django==2.0.0
smtpapi==0.3.1
whitenoise==3.2
